mcp-servers/gcp-vertex-infra/pyproject.toml
[project]
name = "mcp-gcp-vertex"
version = "0.1.0"
description = "MCP server for provisioning GCP Vertex AI projects"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "google-cloud-resource-manager>=1.12.0",
    "google-cloud-aiplatform>=1.50.0",
    "google-cloud-api-keys>=0.5.0",
    "google-cloud-service-usage>=1.9.0",
    "supabase>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "google-auth>=2.0.0",
]

[project.scripts]
mcp-gcp-vertex = "mcp_gcp_vertex.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_gcp_vertex"]
